Status/Return: document expected usage.
The intention for most HIDL users is to declare error codes in HIDL
interfaces and use those. Transport errors are much more rare and in
most cases can just be ignored so that init can restart the process
in question.

This also means that HIDL implementations shouldn't construct the
Status object directly.

Bug: 124861676
Bug: 121004730
Test: N/A
Change-Id: I4a5d79cf453345025c9d0b64c29a4486cc51fab4