Remove unused read/write hidl_version functions.
Originally, these were for hwsm, but the version is encoded in
hidl_string as an FQName now instead. I would try to delete hidl_version
entirely since it doesn't make much sense in HidlSupport.h anymore, but
the number of users means that it is hard to move (it is being used to
help manage interfaces).

It's unlikely that these functions will ever be brought back too. Since
hidl_version can be expressed in the HIDL language, it doesn't make
sense as a primitive in the HIDL language.

Bug: N/A
Test: TH
Change-Id: I09f3920c13dfeab08f562e8d57c47507b06961ed