Check transport before registering.
Note that we already check this when getting a service.

VTS checks that the static VINTF manifest information directly reflects
the HALs that are at runtime. This is a correctness check that ensures
that the manifest is correct. The VINTF manifest needs to be correct so
that:
- the manifest can be read in order to understand what happens at
runtime. For instnace, to understand which HAL interfaces are still
being used (and therefore the cost of deprecating them).
- so that every HAL in the manifest can be certified by VTS

Fixes: 139274536
Test: hidl_test (passes since servers already declare
  TREBLE_TESTING_OVERRIDE in order to load passthrough implementations
  from dlopen(nullptr)).
Test: remove entry from manifest and check for log/that service didn't
  register.

Change-Id: If851e892729cbda5fb4bf014c79242a0ef5b5152