Add helper API to get all HAL instances.
This API is technically already available as defaultServiceManager1_2
listManifestByInterface, but it is somewhat cumbersome to use.

Two things in mind with this change:
a). convenient method for VTS HIDL tests to use instead of having the
VTS test runner invoke a test many times.
b). with an eye towards validating FCM <regex-instance> entries.

Bug: 139437880
Test: updated Lights VTS test.
Change-Id: I2060dcf67a0969ac0099bb4b0f4a021ff6ec6599