getProcessAgeMs: better error messages
There function is jumping to abort on specific boots in two processes
on the device repeatedly. Most likely, it is failing because of ubsan,
but it isn't clear where exactly it is failing, so checking all
assumptions and dumping errors.

Bug: 170456142
Bug: 173264759
Test: adb logcat *:E | grep HidlServiceManagement # no errors on ToT
Test: adb logcat | grep HidlServiceManagement # see registration logs
Merged-In: Ib5c390dd93a9e4f78b42e7172cff74603a89f6cf
Change-Id: Ib5c390dd93a9e4f78b42e7172cff74603a89f6cf